Refactor contact description to use a simple utility class or builder option
