+++ src/test/java/uk/gov/companieshouse/mapper/AttachmentMapperTest.java
@@ -10,7 +10,7 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 import uk.gov.companieshouse.TestData;
 import uk.gov.companieshouse.database.entity.AttachmentEntity;
 import uk.gov.companieshouse.model.Attachment;
-import uk.gov.companieshouse.util.TestUtil;
+import uk.gov.companieshouse.TestUtil;
 
 @ExtendWith(SpringExtension.class)
 public class AttachmentMapperTest {
